калькулятор урок №3

+++ app/src/main/java/YuriusR/calculatorgb/modelView/Calculator.java
@@ -2,5 +2,5 @@ package YuriusR.calculatorgb.modelView;
 
 public interface Calculator {
 
-    float performOperations (float arg1, float arg2,  Operations operations );
+    double performOperations (double arg1, double arg2,  Operations operations );
 }
